refactor: Refactored the reusable location filter for better code organization.
